+++ ModbusSender/PacketModbusSender.cs
@@ -35,13 +35,13 @@ namespace ModbusSender
             }
         }
 
-        public void WriteRegister(ushort devid, ushort index, ushort value)
+        public void WriteRegister(int devid, UInt16 index, UInt16 value)
         {
             lock (lockObj)
             {
                 var cmd = string.Format("MB:{0:X4}:{1:X4}:{2:X4}", devid, index, value);
                 Logger.Instance.Debug(this, "send", cmd);
-		output.SendPacket(cmd);
+		        output.SendPacket(cmd);
             }
         }
     }
